FlaskBackend/requirements.txt
APScheduler==3.11.1
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
ping3==5.1.5
pyasn1==0.6.1
pysnmp==7.1.22
python-dotenv==1.2.1
requests==2.32.5
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
pyzabbix==1.3.1
